refactor(nestjs-oathkeeper): move request interface
